Alpha Themes
This is a first-semester project as a CS beginner. The project is based on HTML and CSS. This project is a prototype for a website based on selling themes. The website contains different sections such as free themes and paid themes and also provides tools to filter between other products and articles. The website is a multipage website and helped me to learn basics about HTML and CSS